Remove __need macros from errno.h (__need_Emath, __need_error_t).
This is fairly complicated, not because the users of __need_Emath and
__need_error_t have complicated requirements, but because the core
changes had a lot of fallout.

__need_Emath is defined only by .S files; what they _really_ need is
for errno.h to avoid declaring anything other than the E-constants
(e.g. 'extern int __errno_location(void);' is a syntax error in
assembly language). This is replaced with a check for __ASSEMBLER__ in
errno.h, plus a carefully documented requirement for bits/errno.h not
to define anything other than macros.  That in turn has the
consequence that bits/errno.h must not define errno - fortunately,
there were only two versions of the definition in a live port, so I
moved them to errno.h and added a new __GLIBC_USE macro to choose
between them.  As a consequence of *that*, attribute_tls_model_ie
needed to move to sys/cdefs.h and be renamed __attribute_tls_model_ie__.

__need_error_t exists for gnulib compatibility in argz.h and argp.h.
error_t itself is a slightly-too-clever Hurdism, an enum containing
all the E-constants, so you can do 'p (error_t) errno' in gdb and get
a symbolic value.  argz.h and argp.h want to fall back to 'int' when
that's not available.  Also, to comply with the rule that bits/errno.h
may only define macros anymore, the definition of error_t had to be
removed from sysdeps/mach/hurd/bits/errno.h somehow.  I have elected
to move error_t to a file generated at build time and make it
available for all supported hosts, not just Hurd.  This allows
creation of a generic bits/types/error_t.h.  However, the price is
that the new definition requires a GCC >=4.5 feature - otherwise it
falls back to 'int'.  And you may regret reading the generated
bits/errno-enum.h.

	* stdlib/errno.h: Remove __need_Emath and __need_error_t logic.
        Include bits/types/error_t.h for error_t.  Declare errno here,
	using either __thread or __errno_location, as directed by
	bits/errno.h.  When __ASSEMBLER__ is defined, provide only the
	E* constants, as macros.  Reorganize file.
	* stdlib/bits/types/error_t.h: New file.
	* include/bits/types/error_t.h: New wrapper.
	* scripts/make-errno-enum.sh: New file.
	* Makeconfig: Add rules to generate bits/errno-enum.h.
	(common-generated): Add bits/errno-enum.h and bits/errno-enum.h.stmp.
	Also add dl-tunable-list.h and dl-tunable-list.stmp, missed in a
	previous patch.
	(dl-tunable-list.h): Generate using a stamp file.
	* stdlib/Makefile: Install bits/types/error_t.h and bits/errno-enum.h.

	* bits/errno.h: Remove logic for __need_Emath.  Document
	requirements for a port-specific bits/errno.h.

	* include/errno.h: Change conditional for exposing internal
	declarations to (not _ISOMAC and not __ASSEMBLER__).
	* include/features.h (__GLIBC_USE_TLS_ERRNO): New feature
	selection macro, defaults to 0.

	* sysdeps/mach/hurd/errnos.awk
	* sysdeps/nacl/errnos.awk
	* sysdeps/unix/sysv/linux/bits/errno.h
	* sysdeps/unix/sysv/linux/alpha/bits/errno.h
	* sysdeps/unix/sysv/linux/hppa/bits/errno.h
	* sysdeps/unix/sysv/linux/mips/bits/errno.h
	* sysdeps/unix/sysv/linux/sparc/bits/errno.h:
	Add multiple-include guard and check against improper inclusion.
	Remove __need_Emath logic. Don't declare errno here. Ensure all
	constants are defined as simple integer literals.  Redefine
	__GLIBC_USE_TLS_ERRNO to 1 if appropriate.  Consistent formatting.
	* sysdeps/mach/hurd/bits/errno.h: Regenerate.

	* argp/argp.h, string/argz.h: Include bits/types/error_t.h for error_t.

	* sysdeps/i386/i686/fpu/multiarch/s_cosf-sse2.S
	* sysdeps/i386/i686/fpu/multiarch/s_sincosf-sse2.S
	* sysdeps/i386/i686/fpu/multiarch/s_sinf-sse2.S
	* sysdeps/x86_64/fpu/s_cosf.S
	* sysdeps/x86_64/fpu/s_sincosf.S
	* sysdeps/x86_64/fpu/s_sinf.S
	Just include errno.h; don't define __need_Emath or include
	bits/errno.h directly.

	* include/libc-symbols.h (attribute_tls_model_ie): Remove.
	* misc/sys/cdefs.h (__attribute_tls_model_ie__): New.
	* elf/dl-error-skeleton.c, include/errno.h, include/netdb.h
	* include/resolv.h, locale/localeinfo.h, malloc/arena.c
	* sunrpc/rpc_thread.c, sysdeps/generic/libc-tsd.h
	* sysdeps/ieee754/ldbl-opt/math_ldbl_opt.c
	* sysdeps/ieee754/ldbl-opt/math_ldbl_opt.h
	* sysdeps/powerpc/nofpu/soft-supp.h
	* sysdeps/powerpc/soft-fp/sfp-machine.h
	Use __attribute_tls_model_ie__ and/or __attribute_const__.